+++ doc/changelog.md
@@ -13,6 +13,7 @@ Jump to:
 
 Description
 
+- RequestBatch rewrite
 - Fix dragon build logging bug
 - Merge core refactor into MLI feature branch
 - Implement asynchronous notifications for shared data